telegram-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-telegram-bot==20.7
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
aiohttp==3.9.1